history: fold into one module, move routes to routes.py, drop comments
`history_recorder.py` and `history_routes.py` are gone. Everything that is
not a route now lives in `gradio/history.py`; the four route handlers moved
into `routes.py` alongside the rest of the app's routes, under the existing
`run_history` guard.

`init_history_state` no longer allocates `bucket_history_cache` and its
lock, which have been dead since `store_for` was removed.

Inline comments are gone and docstrings are one line each.
